nimusuggest v3
Rework `nimsuggest` to use caching to make usage of ide commands more efficient.
Previously, all commands no matter what the state of the process is were causing
clean build. In the context of Language Server Protocol(LSP) and lsp clients
this was causing perf issues and overall instability.

- added two new commands: `recompile` to do clean compilation and `changed`
which can be used by the IDEs to notify that a particular file has been changed.
The later can be utilized when using LSP file watches.